add 'import js as thing' tests
