--- test/e2e/turbopack-import-type-bytes/app/api/route.js
@@ -1,12 +1,0 @@
-import bytes from './data.bin' with { type: 'bytes' }
-
-export async function GET(_req) {
-  return Response.json(
-    {
-      instanceofUint8Array: bytes instanceof Uint8Array,
-      length: bytes.length,
-      content: new TextDecoder().decode(bytes),
-    },
-    { status: 200 }
-  )
-}
