--- test/e2e/turbopack-import-type-text/next.config.ts
@@ -1,5 +1,0 @@
-export default {
-  experimental: {
-    turbopackImportTypeText: true,
-  },
-}
