+++ test/e2e/turbopack-import-with-type/app/api/data.txt
