+++ test/e2e/turbopack-import-with-type/app/api/some.js
@@ -1,0 +1,3 @@
+export const nope = 'nope'
+
+throw new Error('please dont execute me')
