+++ test/e2e/turbopack-import-with-type/index.test.ts
@@ -1,0 +1,42 @@
+import { nextTestSetup } from 'e2e-utils'
+
+const jsContent = `export const nope = 'nope'
+
+throw new Error('please dont execute me')
+`
+
+describe('turbopack-import-with-type', () => {
+  const { next, skipped } = nextTestSetup({
+    files: __dirname,
+    skipDeployment: true,
+  })
+
+  if (skipped) {
+    return
+  }
+
+  // Testing this together on one route ensures we also avoid weird duplicate module ident things
+  it('supports import with type: text and type: bytes', async () => {
+    const response = JSON.parse(await next.render('/api'))
+    expect(response).toEqual({
+      text: {
+        typeofString: true,
+        length: 12,
+        content: 'hello world\n',
+      },
+      jsAsText: {
+        typeofString: true,
+        content: jsContent,
+      },
+      bytes: {
+        instanceofUint8Array: true,
+        length: 18,
+        content: 'this is some data\n',
+      },
+      jsAsBytes: {
+        instanceofUint8Array: true,
+        content: jsContent,
+      },
+    })
+  })
+})
