+++ test/e2e/turbopack-import-with-type/next.config.ts
@@ -1,5 +1,6 @@
 export default {
   experimental: {
     turbopackImportTypeBytes: true,
+    turbopackImportTypeText: true,
   },
 }
